Bugfix/tests catch (#309)
* fixed empty catch blocks in tests

* tests code analysis fixes

* bugfix IfMatch graph tests

* bugfix assumeTrue to skip tests

+++ src/test/java/com/arangodb/example/ExampleBase.java
@@ -25,39 +25,35 @@ import org.junit.BeforeClass;
 
 import com.arangodb.ArangoCollection;
 import com.arangodb.ArangoDB;
-import com.arangodb.ArangoDBException;
 import com.arangodb.ArangoDatabase;
 
 /**
  * @author Mark Vollmary
- *
  */
 public class ExampleBase {
 
-	protected static final String DB_NAME = "json_example_db";
-	protected static final String COLLECTION_NAME = "json_example_collection";
+    private static final String DB_NAME = "json_example_db";
+    protected static final String COLLECTION_NAME = "json_example_collection";
 
-	protected static ArangoDB arangoDB;
-	protected static ArangoDatabase db;
-	protected static ArangoCollection collection;
+    private static ArangoDB arangoDB;
+    protected static ArangoDatabase db;
+    protected static ArangoCollection collection;
 
-	@BeforeClass
-	public static void setUp() {
-		arangoDB = new ArangoDB.Builder().build();
-		try {
-			arangoDB.db(DB_NAME).drop();
-		} catch (final ArangoDBException e) {
-		}
-		arangoDB.createDatabase(DB_NAME);
-		db = arangoDB.db(DB_NAME);
-		db.createCollection(COLLECTION_NAME);
-		collection = db.collection(COLLECTION_NAME);
-	}
+    @BeforeClass
+    public static void setUp() {
+        arangoDB = new ArangoDB.Builder().build();
+        if (arangoDB.db(DB_NAME).exists())
+            arangoDB.db(DB_NAME).drop();
+        arangoDB.createDatabase(DB_NAME);
+        db = arangoDB.db(DB_NAME);
+        db.createCollection(COLLECTION_NAME);
+        collection = db.collection(COLLECTION_NAME);
+    }
 
-	@AfterClass
-	public static void tearDown() {
-		db.drop();
-		arangoDB.shutdown();
-	}
+    @AfterClass
+    public static void tearDown() {
+        db.drop();
+        arangoDB.shutdown();
+    }
 
 }
