+++ src/test/java/com/arangodb/serde/CustomSerdeTest.java
@@ -47,7 +47,7 @@ import static org.junit.Assert.assertThat;
  */
 public class CustomSerdeTest {
 
-    private static String COLLECTION_NAME = "collection";
+    private static final String COLLECTION_NAME = "collection";
 
     private ArangoDatabase db;
     private ArangoCollection collection;
@@ -75,7 +75,8 @@ public class CustomSerdeTest {
 
     @After
     public void shutdown() {
-        db.drop();
+        if (db.exists())
+            db.drop();
     }
 
     @Test
